requirements.txt
streamlit
pandas
bokeh==2.4.3
space-classy @ git+https://github.com/maxmahlke/classy@main
space-rocks
altair<5
tf-keras
